Safari-609.1.11.1/LayoutTests/js/kde/const-expected.txt

KDE JS Test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS c is eleven
PASS cant assign to const
PASS c is eleven
PASS v is 1
PASS successfullyParsed is true

TEST COMPLETE
